Disable auto transfer mode
Prior to this change, "show file" shows, among other things:

 Transfer mode:           automatic
 File patterns:           automatic (SHOW PATTERNS for list)
 Default file type:       binary

With this change:

 Transfer mode:           manual
 File patterns:           automatic (but disabled by TRANSFER-MODE MANUAL)
 File type:               binary

The modern assumption is a byte-accurate transfer of files.  We have had
a proliferation of file types, extensions, and complicating
circumstances since the earlier days of Kermit.

By changing this default, we disable the heuristic for attempting to
guess the type of files, and convert the existing binary default into a
binary setting.

This can always be changed by the user, but the idea is to not violate
the principle of least surprise.  If the user asks to transfer a file,
we assume the user wants an exact transfer of the file unless stated
otherwise.

Additionally, some platforms (eg, HP48 calculators) have wildly
different behavior depending on whether a text or binary transfer is
requested.  By defaulting to manual mode, the user is in charge and
further surprises that may be caused by "set file type" being ignored
can be avoided.