Eliminate use of 'volatile' for OS2 (!NT) signal handlers
During the development cycle leading up to the C-Kermit 6.0 release
the signal handler function pointers were designated 'volatile'.

There is no obvious reason for the use of 'volatile' for function
pointers which are not shared by multiple threads.  One possibility
is that the Visual Age compiler used at the time declared signal
handler pointer types to be volatile.  This is not true for the
Open Watcom compiler used to build C-Kermit for OS/2 today.

This change removes the 'volatile' designation as a precursor to
subsequent change which will consolidate signal handler function
pointer type declarations.